# openapi.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "DICOMweb",
      "version": "2023a"
    },
...
    "paths": {
...
      "/studies": {
        "get": {
          "tags": [
            "QIDO-RS"
          ],
          "summary": "Search for all Studies",
          "operationId": "SearchForAllStudies",
          "parameters": [
            {
              "$ref": "#/components/parameters/search-accept-hp"
            },
            {
              "$ref": "#/components/parameters/charset-hp"
            },
            {
              "$ref": "#/components/parameters/search-accept-qp"
            },
            {
              "$ref": "#/components/parameters/charset-qp"
            },
            {
              "$ref": "#/components/parameters/match"
            },
            {
              "$ref": "#/components/parameters/fuzzymatching"
            },
            {
              "$ref": "#/components/parameters/includefield"
            },
            {
              "$ref": "#/components/parameters/limit"
            },
            {
              "$ref": "#/components/parameters/offset"
            }
          ],
          "responses": {
            "200": {
              "$ref": "#/components/responses/searchResults"
            },
            "204": {
              "description": "The search completed successfully, but there were zero results."
            },
            "400": {
              "description": "There was a problem with the request."
            },
            "413": {
              "description": "The search was too broad, and the body of the response should contain a Status Report with additional information about the failure."
            }
          }
        },
        "post": {
          "tags": [
            "STOW-RS"
          ],
          "summary": "Store Studies",
          "operationId": "StoreStudies",
          "parameters": [
            {
              "$ref": "#/components/parameters/search-accept-hp"
            },
            {
              "$ref": "#/components/parameters/search-accept-qp"
            }
          ],
          "requestBody": {
            "$ref": "#/components/requestBodies/storeStudies"
          },
          "responses": {
            "200": {
              "$ref": "#/components/responses/storeInstancesSuccessful"
            },
            "202": {
              "$ref": "#/components/responses/storeInstancesWarnings"
            },
            "400": {
              "description": "The origin server was unable to store any instances due to bad syntax."
            },
            "409": {
              "$ref": "#/components/responses/storeInstancesConflict"
            },
            "415": {
              "description": "The origin server does not support the media type specified in the Content-Type header field of the request."
            }
          }
        }
      },
      "/studies/{study}": {
        "get": {
          "tags": [
            "WADO-RS"
          ],
          "summary": "Retrieve Study",
          "operationId": "RetrieveStudy",
          "parameters": [
            {
              "$ref": "#/components/parameters/retrieve-accept-hp"
            },
            {
              "$ref": "#/components/parameters/retrieve-accept-qp"
            },
            {
              "$ref": "#/components/parameters/study"
            }
          ],
          "responses": {
            "200": {
              "$ref": "#/components/responses/retrievedInstances"
            },
            "206": {
              "$ref": "#/components/responses/retrievedInstancesPartial"
            },
            "400": {
              "description": "The origin cannot process the request because of errors in the request headers or parameters."
            },
            "404": {
              "description": "The Study does not exist."
            },
            "406": {
              "description": "The origin server does not support any of the Acceptable Media Types."
            },
            "410": {
              "description": "The Study has been deleted."
            },
            "413": {
              "description": "The Study is too large to be returned by the origin server."
            }
          }
        },
        "post": {
          "tags": [
            "STOW-RS"
          ],
          "summary": "Store Study",
          "operationId": "StoreStudy",
          "parameters": [
            {
              "$ref": "#/components/parameters/search-accept-hp"
            },
            {
              "$ref": "#/components/parameters/search-accept-qp"
            },
            {
              "name": "study",
              "in": "path",
              "description": "The Study Instance UID of Study.",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "$ref": "#/components/requestBodies/storeStudies"
          },
          "responses": {
            "200": {
              "$ref": "#/components/responses/storeInstancesSuccessful"
            },
            "202": {
              "$ref": "#/components/responses/storeInstancesWarnings"
            },
            "400": {
              "description": "The origin server was unable to store any instances due to bad syntax."
            },
            "409": {
              "$ref": "#/components/responses/storeInstancesConflict"
            },
            "415": {
              "description": "The origin server does not support the media type specified in the Content-Type header field of the request."
            }
          }
        }
      },
...
    },
    "components": {
      "parameters": {
        "charset-qp": {
          "name": "charset",
          "in": "query",
          "description": "Comma-separated list of Acceptable Character Sets of the response payload.",
          "schema": {
            "type": "string"
          }
        },
        "charset-hp": {
          "name": "Accept-Charset",
          "in": "header",
          "description": "The Acceptable Character Sets of the response payload.",
          "schema": {
            "type": "string"
          }
        },
        "search-accept-qp": {
          "name": "accept",
          "in": "query",
          "description": "The Acceptable Media Types for the response payload.",
          "schema": {
            "$ref": "#/components/schemas/search-media-type"
          }
        },
        "search-accept-hp": {
          "name": "Accept",
          "in": "header",
          "description": "The Acceptable Media Types for the response payload.",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/search-media-type-wc"
          }
        },
...
        "retrieve-accept-qp": {
          "name": "accept",
          "in": "query",
          "description": "The Acceptable Media Types for the response payload.",
          "schema": {
            "$ref": "#/components/schemas/retrieve-media-type"
          }
        },
        "retrieve-accept-hp": {
          "name": "Accept",
          "in": "header",
          "description": "The Acceptable Media Types for the response payload.",
          "required": true,
          "schema": {
            "$ref": "#/components/schemas/retrieve-media-type-wc"
          }
        },
...
        "study": {
          "name": "study",
          "in": "path",
          "description": "The Study Instance UID of a Study managed by the Studies Service.",
          "required": true,
          "schema": {
            "type": "string"
          }
        },
...
        "match": {
          "name": "match",
          "in": "query",
          "description": "One or more DICOM Attribute/Values pairs specify the matching criteria for the search.",
          "schema": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "fuzzymatching": {
          "name": "fuzzymatching",
          "in": "query",
          "description": "A single parameter specifies whether Fuzzy Matching of Person Names is to be performed.",
          "schema": {
            "type": "boolean",
            "default": false
          }
        },
        "includefield": {
          "name": "includefield",
          "in": "query",
          "description": "A parameter specifies the Attributes that should be included in the response. The value is either a comma-separated list of attributes, or the single keyword \"all\", which means that all available attributes of the object should be included in the response.",
          "schema": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "limit": {
          "name": "limit",
          "in": "query",
          "description": "A single parameter specifies the maximum number of matches the origin server shall return in a single response.",
          "schema": {
            "type": "integer"
          }
        },
        "offset": {
          "name": "offset",
          "in": "query",
          "description": "A single parameter specifies the number of matches the origin server shall skip before the first returned match.",
          "schema": {
            "type": "integer",
            "default": 0
          }
        },
...
      },
      "schemas": {
...
        "search-media-type": {
          "type": "string",
          "enum": [
            "multipart/related;type=\"application/dicom+xml\"",
            "application/dicom+json"
          ]
        },
        "search-media-type-wc": {
          "type": "string",
          "enum": [
            "*/*",
            "multipart/related;type=\"application/dicom+xml\"",
            "application/dicom+json"
          ]
        },
...
        "retrieve-media-type": {
          "type": "string",
          "enum": [
            "multipart/related;type=\"application/dicom\"",
            "multipart/related;type=\"application/octet-stream\"",
            "multipart/related;type=\"image/jpeg\"",
            "multipart/related;type=\"image/jls\"",
            "multipart/related;type=\"image/jp2\"",
            "multipart/related;type=\"image/jpx\"",
            "multipart/related;type=\"image/dicom-rle\"",
            "multipart/related;type=\"video/mpeg\"",
            "multipart/related;type=\"video/mp4\"",
            "multipart/related;type=\"video/H265\"",
            "multipart/related;type=\"text/xml\"",
            "multipart/related;type=\"application/pdf\"",
            "application/dicom"
          ]
        },
        "retrieve-media-type-wc": {
          "type": "string",
          "enum": [
            "*/*",
            "multipart/related;type=\"application/dicom\"",
            "multipart/related;type=\"application/octet-stream\"",
            "multipart/related;type=\"image/jpeg\"",
            "multipart/related;type=\"image/jls\"",
            "multipart/related;type=\"image/jp2\"",
            "multipart/related;type=\"image/jpx\"",
            "multipart/related;type=\"image/dicom-rle\"",
            "multipart/related;type=\"video/mpeg\"",
            "multipart/related;type=\"video/mp4\"",
            "multipart/related;type=\"video/H265\"",
            "multipart/related;type=\"text/xml\"",
            "multipart/related;type=\"application/pdf\"",
            "application/dicom"
          ]
        },
...
        "dicom-json": {
          "title": "DICOM JSON Model",
          "externalDocs": {
            "description": "DICOM PS3.18 F",
            "url": "https://dicom.nema.org/medical/dicom/current/output/html/part18.html#chapter_F"
          },
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "{ggggeeee}": {
                "required": [
                  "vr"
                ],
                "properties": {
                  "vr" : {
                    "type": "string",
                    "enum": [
                      "AE",
                      "AS",
                      "AT",
                      "CS",
                      "DA",
                      "DS",
                      "DT",
                      "FD",
                      "FL",
                      "IS",
                      "LO",
                      "LT",
                      "OB",
                      "OD",
                      "OF",
                      "OL",
                      "OV",
                      "OW",
                      "PN",
                      "SH",
                      "SL",
                      "SQ",
                      "SS",
                      "ST",
                      "SV",
                      "TM",
                      "UC",
                      "UI",
                      "UL",
                      "UN",
                      "UR",
                      "US",
                      "UT",
                      "UV"
                    ]
                  },
                  "Value": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "requestBodies": {
        "storeStudies": {
          "content": {
            "multipart/related;type=application/dicom": {},
            "multipart/related;type=application/dicom+xml": {},
            "multipart/related;type=application/dicom+json": {}
          }
        }
      },
      "responses": {
...
        "searchResults": {
          "description": "The search completed successfully, and the results are contained in the payload",
          "content": {
            "multipart/related;type=application/dicom+xml": {},
            "application/dicom+json": {
              "schema": {
                "$ref": "#/components/schemas/dicom-json"
              }
            }
          }
        },
        "storeInstancesSuccessful": {
          "description": "The origin server successfully stored all Instances.",
          "content": {
            "application/dicom+xml": {},
            "application/dicom+json": {
              "schema": {
                "$ref": "#/components/schemas/dicom-json"
              }
            }
          }
        },
        "storeInstancesWarnings": {
          "description": "The origin server stored some of the Instances but warnings or failures exist for others.",
          "content": {
            "application/dicom+xml": {},
            "application/dicom+json": {
              "schema": {
                "$ref": "#/components/schemas/dicom-json"
              }
            }
          }
        },
        "storeInstancesConflict": {
          "description": "The request was formed correctly but the origin server was unable to store any instances due to a conflict in the request.",
          "content": {
            "application/dicom+xml": {},
            "application/dicom+json": {
              "schema": {
                "$ref": "#/components/schemas/dicom-json"
              }
            }
          }
        },
        "retrievedInstances": {
          "description": "The response payload contains representations for all of the Target Resource(s).",
          "content": {
            "multipart/related;type=\"application/dicom\"": {},
            "multipart/related;type=\"application/octet-stream\"": {},
            "multipart/related;type=\"application/pdf\"": {},
            "multipart/related;type=\"image/jpeg\"": {},
            "multipart/related;type=\"image/jls\"": {},
            "multipart/related;type=\"image/jp2\"": {},
            "multipart/related;type=\"image/jpx\"": {},
            "multipart/related;type=\"image/dicom-rle\"": {},
            "multipart/related;type=\"text/xml\"": {},
            "multipart/related;type=\"video/mpeg\"": {},
            "multipart/related;type=\"video/mp4\"": {},
            "multipart/related;type=\"video/H265\"": {},
            "application/dicom": {}
          }
        },
        "retrievedInstancesPartial": {
          "description": "The response payload contains representations for some, but not all, of the Target Resource(s).",
          "content": {
            "multipart/related;type=\"application/dicom\"": {},
            "multipart/related;type=\"application/octet-stream\"": {},
            "multipart/related;type=\"application/pdf\"": {},
            "multipart/related;type=\"image/jpeg\"": {},
            "multipart/related;type=\"image/jls\"": {},
            "multipart/related;type=\"image/jp2\"": {},
            "multipart/related;type=\"image/jpx\"": {},
            "multipart/related;type=\"image/dicom-rle\"": {},
            "multipart/related;type=\"text/xml\"": {},
            "multipart/related;type=\"video/mpeg\"": {},
            "multipart/related;type=\"video/mp4\"": {},
            "multipart/related;type=\"video/H265\"": {},
            "application/dicom": {}
          }
        },
...
      }
    }
  }